refactor: Extract logic from StatsController
Extracted the business logic from `StatsController::index` into a dedicated action class `FetchStatsOverviewAction`. This improves code readability and maintainability by adhering to the single responsibility principle.

- Created `App\Actions\Stats\FetchStatsOverviewAction`
- Refactored `App\Http\Controllers\StatsController` to use the new action
- Removed `parsePeriod` and `getFilteredExercises` methods from the controller as they are now handled by the action
- Ensured PSR-12 compliance and strict typing
