+++ app/Http/Controllers/StatsController.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace App\Http\Controllers;
 
+use App\Actions\Stats\FetchStatsOverviewAction;
 use App\Models\Exercise;
 use App\Services\StatsService;
 use Inertia\Inertia;
@@ -17,28 +18,12 @@ class StatsController extends Controller
     /**
      * Display the main statistics dashboard.
      */
-    public function index(\Illuminate\Http\Request $request): \Inertia\Response
+    public function index(\Illuminate\Http\Request $request, FetchStatsOverviewAction $fetchStatsOverview): \Inertia\Response
     {
-        $user = $this->user();
-        $days = $this->parsePeriod($request->query('period', '30j'));
-
-        // Body metrics and weight history
-        $bodyMetrics = $this->statsService->getLatestBodyMetrics($user);
-        $weightHistory = $this->statsService->getWeightHistory($user, $days);
-
-        return Inertia::render('Stats/Index', [
-            'volumeTrend' => $this->statsService->getVolumeTrend($user, $days),
-            'muscleDistribution' => $this->statsService->getMuscleDistribution($user, $days),
-            'monthlyComparison' => $this->statsService->getMonthlyVolumeComparison($user),
-            'weightHistory' => $weightHistory,
-            'bodyFatHistory' => $this->statsService->getBodyFatHistory($user, $days),
-            'durationHistory' => $this->statsService->getDurationHistory($user, 30),
-            'latestWeight' => $bodyMetrics['latest_weight'],
-            'weightChange' => $bodyMetrics['weight_change'],
-            'bodyFat' => $bodyMetrics['latest_body_fat'],
-            'exercises' => $this->getFilteredExercises($user->id),
-            'selectedPeriod' => $request->query('period', '30j'),
-        ]);
+        return Inertia::render('Stats/Index', $fetchStatsOverview->execute(
+            $this->user(),
+            $request->query('period', '30j')
+        ));
     }
 
     /**
@@ -49,22 +34,5 @@ class StatsController extends Controller
         return response()->json([
             'progress' => $this->statsService->getExercise1RMProgress($this->user(), $exercise->id),
         ]);
-    }
-
-    protected function parsePeriod(string $period): int
-    {
-        return match ($period) {
-            '7j' => 7,
-            '30j' => 30,
-            '90j' => 90,
-            '1a' => 365,
-            default => 30,
-        };
-    }
-
-    /** @return \Illuminate\Database\Eloquent\Collection<int, Exercise> */
-    private function getFilteredExercises(int $userId): \Illuminate\Database\Eloquent\Collection
-    {
-        return Exercise::getCachedForUser($userId);
     }
 }
